Move authorize REST call to after EAP (#4)
We want to see if the eap authentication succeeds before calling out to
our backend API.  The way EAP-PEAP works is that it will call our API
multiple (3) times to do the authentication.  By moving this rest call to
below the PEAP authorization we will reduce the number of calls to 2.

Getting this down to 1 is a future task, which requires changes to the
frontend configuration which is quite complex.