Add FreeRadius Log Exporter To Docker Containers
We are adding a Prometheus instance to Govwifi to help
us collect better metrics. In order to do this we need
to add the FreeRADIUS exporter service and it's supporting
packages to the Docker containers that host our RADIUS service.

You can find more information about the FreeRADIUS exporter here:
https://github.com/bvantagelimited/freeradius_exporter