Download the freeradius_exporter in the Dockerfile
Rather than using the binary that's committed to the Git repository.

Committing binaries to Git repositories is generally something to
avoid, mostly because it inflates the size of the repository. In this
specific case, I'm also not sure where the binary came from, and what
version/code it corresponds to.

Fix that by just downloading a binary in the Dockerfile. The version
I've chosen is just the latest release (it also seems to differ from
the current binary). Include a checksum check to ensure what the
binary doesn't change unexpectedly.

This commit also removes the freeradius_exporter binary.