Use Ruby-3.0.3-alpine image
The Health Check uses Ruby 3.0.3 so the docker image
needs to be updated. The image uses freeradius-rest-3.0.25-r1
which includes a fix to the issue of the validation of Rest
authentication responses breaking.

As this image already includes Ruby, it doesn't need to be
fetched and installed and so this portion can be removed.