Assert the freeradius version in the Dockerfile
Dockerfiles are awful, especially for composing different bits of
software. Changing the base image can affect the freeradius version
used.

So that changes to the freeradius version is intentional, add a
constraint to the freeradius package version in the Dockerfile.

The recent breaking change occurred between what I would consider
patch versions (3.0.17 to 3.0.25), so make the version quite specific.