Add test suite.
Add a new Sinatra-based test-app which calls various combinations
of radclient/eapol_test to ensure that authentication requests are
accepted/rejected as expected.

Ensure that the FreeRADIUS server starts correctly.

Re-enable the linting of the healthcheck app, which was disabled in
bb870d36424cd48a9b08a163ab7805153e89fdf7 ('Enable the FreeRADIUS
server to be run locally') when it was necessary to remove test/lint
calls from the Makefile for deployment.