Fix critical issue with radius docker image
The Dockerfile.raddb that we use to create the ECR image that our radius tasks
boot from, was erroring when unzipping this package:
awscli-bundle/packages/setup/wheel-0.33.6.tar.gz

Error occurred whenever a build of the docker image was attempted. This was
caused by an outdated version of the aws-cli. I have updated the aws-cli to
version 2, which has solved the issue.  Installing with curl/wget was
cumbersome, so this has been changed to use the official alpine version of
aws-cli with the apk add command instead.