Use X509_V_FLAG_PARTIAL_CHAIN preprocessor flag.
When compiling FreeRadius, we set the X509_V_FLAG_PARTIAL_CHAIN
preprocessor flag. This causes intermediate certificates in the
trust store to be treated as trust-anchors, in the same way as
self-signed root CA certificates.

So if we want to limit the certificates that are accepted to ones
that have been issued by a specific intermediate certificate, we
need to make sure that that certificate is the only one in the
trust store.

In the tests we therefore need to remove the root certificate from
the trust store so we don't accept client certificates issued by
other branches in the PKI.